Issue542 Move notif dialog out of MusicController, use startup task
