+++ core/MusicControl/pom.xml
@@ -115,11 +115,6 @@
             <artifactId>org-openide-util</artifactId> 
             <version>${netbeans.version}</version>  
         </dependency>
-        <dependency> 
-            <groupId>org.netbeans.api</groupId> 
-            <artifactId>org-openide-windows</artifactId> 
-            <version>${netbeans.version}</version>  
-        </dependency>
         <dependency>
             <groupId>org.netbeans.api</groupId> 
             <artifactId>org-openide-util-lookup</artifactId>
